+++ packages/vm/src/middleware/deterministic.rs
@@ -165,7 +165,7 @@ fn parse_wasm_opcode(opcode: &Operator) -> Result<(), CompileError> {
 /// https://github.com/CosmWasm/cosmwasm/issues/311
 #[cfg(all(test, feature = "default-singlepass"))]
 mod tests {
-    use super::*;
+    // No 'use super::*;' here. This is strange and means we are not testing the functions in this module directly.
     use crate::backends::compile;
     use crate::errors::VmError;
     use wabt::wat2wasm;
@@ -200,22 +200,12 @@ mod tests {
                     f32.convert_u/i32
                 ))
             "#;
-
         let wasm = wat2wasm(input).unwrap();
         let res = compile(&wasm);
 
-        let failure = res.err().expect("compile should have failed");
-
-        if let VmError::CompileErr { source, .. } = &failure {
-            if let CompileError::InternalError { msg } = source {
-                assert_eq!(
-                    "Codegen(\"ValidationError { msg: \\\"non-deterministic opcode: F32ConvertI32U\\\" }\")",
-                    msg.as_str()
-                );
-                return;
-            }
+        match res.err().unwrap() {
+            VmError::CompileErr { msg, .. } => assert_eq!(msg, "Compile error: InternalError { msg: \"Codegen(\\\"ValidationError { msg: \\\\\\\"non-deterministic opcode: F32ConvertI32U\\\\\\\" }\\\")\" }"),
+            e => panic!("unexpected error: {:?}", e),
         }
-
-        panic!("unexpected error: {:?}", failure)
     }
 }